Implemented network usage overview

+++ GW3/GWLogger/Live/GatewayInformation.cs
@@ -13,6 +13,6 @@
         public string RunningTime { get; set; }
         public int? NbClients { get; set; }
         public int? NbServers { get; set; }
-        public int? Network { get; set; }
+        public double? Network { get; set; }
     }
 }
